requirements.txt
Flask
pycrypto
python-firebase
urllib3